+++ core/src/com/unciv/logic/files/FileChooser.kt
@@ -20,7 +20,7 @@ import com.badlogic.gdx.utils.Array as GdxArray
 import com.unciv.Constants
 import com.unciv.models.UncivSound
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.AutoScrollPane
+import com.unciv.ui.components.widgets.AutoScrollPane
 import com.unciv.ui.components.input.KeyboardBinding
 import com.unciv.ui.components.UncivTextField
 import com.unciv.ui.components.extensions.addSeparator
