+++ core/src/com/unciv/ui/crashhandling/CrashScreen.kt
@@ -10,7 +10,7 @@ import com.unciv.Constants
 import com.unciv.UncivGame
 import com.unciv.logic.files.UncivFiles
 import com.unciv.models.ruleset.RulesetCache
-import com.unciv.ui.components.AutoScrollPane
+import com.unciv.ui.components.widgets.AutoScrollPane
 import com.unciv.ui.components.extensions.addBorder
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.components.extensions.setFontSize
