+++ core/src/com/unciv/ui/screens/mapeditorscreen/tabs/MapEditorSaveTab.kt
@@ -9,7 +9,7 @@ import com.unciv.logic.files.MapSaver
 import com.unciv.logic.map.MapGeneratedMainType
 import com.unciv.logic.map.TileMap
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.AutoScrollPane
+import com.unciv.ui.components.widgets.AutoScrollPane
 import com.unciv.ui.components.TabbedPager
 import com.unciv.ui.components.UncivTextField
 import com.unciv.ui.components.extensions.isEnabled
