+++ core/src/com/unciv/ui/screens/modmanager/ModManagementScreen.kt
@@ -17,7 +17,7 @@ import com.unciv.models.ruleset.Ruleset
 import com.unciv.models.ruleset.RulesetCache
 import com.unciv.models.tilesets.TileSetCache
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.AutoScrollPane
+import com.unciv.ui.components.widgets.AutoScrollPane
 import com.unciv.ui.components.ExpanderTab
 import com.unciv.ui.components.UncivTextField
 import com.unciv.ui.components.WrappableLabel
