+++ core/src/com/unciv/ui/screens/worldscreen/NotificationsScroll.kt
@@ -13,7 +13,7 @@ import com.badlogic.gdx.utils.Align
 import com.unciv.GUI
 import com.unciv.logic.civilization.Notification
 import com.unciv.logic.civilization.NotificationCategory
-import com.unciv.ui.components.AutoScrollPane as ScrollPane
+import com.unciv.ui.components.widgets.AutoScrollPane as ScrollPane
 import com.unciv.ui.components.ColorMarkupLabel
 import com.unciv.ui.components.WrappableLabel
 import com.unciv.ui.components.input.onClick
